claimSubmission/server/requirements.txt
Flask==1.1.2
pika==1.1.0
python-dotenv==0.12.0
Flask-Cors==3.0.8
requests==2.21.0
dialogflow==0.8.0
protobuf==3.11.3
pybase64==1.0.1